output/dnssim: refactor - move extra dnssim headers
