+++ src/output/dnssim/internal.h
